Add GH CI action for Rails
Uses the provided template with an updated PG and test script.

The template was using an older version of the test action which doesn't
support the latest Ruby 3.1.1. This updates the action version to use
the latest so we can always have patch releases in a faster time frame.

The security linters are frequently updated. For CI we want to always
have them at their best so we do not include them in the project's
Gemfile. Instead we install them directly on CI when we need to run
them.

Per the action definition, this updates the database config to read a DB
URL when present. The PG config will use the url over any other settings
so it will work on CI. On dev systems, which should not have that set,
it will use the defaults per the config file.