Fix annotation config
Accidentally set the `frozen` option to `true` instead of `false`.
Annotate won't update the docs when it is run with `frozen=true`; and we
want the schema comments to update.

We generally write our comments with YARD. However, the yard format for
annotate doesn't include the various null constraints and defaults for
the columns. That is important information so we are going to use the
markdown format which should generate a nice table in docs.