Add basic User model
We are omitting an ADR for this as this is just a demo app. Normally the
thought process that goes into a user's attributes is long and complex.
In our case, it's highly probable the user's attributes are defined by
another account / authentication system. If that were the case we'd copy
the ADR from there into here as the documentation.

In lieu of an ADR, here are the rough thoughts going to the user model:

- A user's login may not be a handle (e.g. an email) but we want our
  system to be able to have a "short unique reference" for each user.

  As we want to use this for a reference we are capping it at 120
  characters (not bytes, as we want to support unicode characters).
  Even requiring these to be unique via the "lower case" interpretation,
  we shouldn't run out of unique names as there are trillions of
  possible values.

- We need a display name for the user

  Names are highly subjective and may take different forms for different
  purposes. This is simply a display name for us so we want to be as
  flexible to the user as possible. But we want to have some cap on the
  length to be respectful for UI space.

  See also https://www.kalzumeus.com/2010/06/17/falsehoods-programmers-believe-about-names/

As a minor performance optimization, we do not run the extra DB unique
checks when the handle has not changed. Rails will not send the `handle`
data in an `UPDATE` command if it has not been modified. Thus there's no
reason for us to increase latency doing an unnecessary DB round trip.