Propose data model for initial message system
This is still a proposal and will be discussed with the team before
finalization.

This focuses on the data model for the direct messages feature. However,
it takes into account the road map for the proposed design. While we are
not going to be implementing anything beyond the direct messages we want
to be aware of where we are going so we don't make a design choice that
hinders that future progress. Or worse requires a lot of work to
redesign it all and deal with risky data migrations.