Add basic depot model
This creates the `depots` table per ADR 20220310190424.

While not called out explicitly in the text of the ADR, the association
between a `Depot` and other resources is one-to-one per the schema
diagram. Thus, to preserve the data integrity of those associations we
make the index for the `receiver` reference unique. That ensures we only
ever have one such relation in the database.