Make direct message depots first class
This is utilizing both polymorphic associations and STI. Utilizing STI
allows us to setup some common behavior in the parent class and then
customize things as necessary in the child class. Further, it can allow
us to setup associations directly with a child class. Those types of
associations are more explicit and provide a build-in way of specifying
a default scope without actually writing default scopes; Rails default
scopes can get tricky and if we can avoid them via first class
associations that is ideal.