Add direct depot association to user
Due to time constraints we're punting on the unit tests. This is fairly
boilerplate Rails associations so we have a high degree of confidence it
works. However, this is still contractual behavior of our class, so it
should be tested. Anecdotally, these types of tests have caught
regressions during upgrades so there is additional value in them.