Add basic message model
This creates the `messages` table per ADR 20220310190424.

Per the ADR we establish domain language association names and not the
Rails expected defaults (e.g. `author` and `author_id` instead of `user`
and `user_id` respectfully).

Data integrity is important so we include basic validations and specs
for that behavior. However, due to time constraints we are punting on
association unit tests.