Establish base API v1 controller
As this is a JSON API we don't need all of the features provided by
`ActionController::Base`. Rails provides a JSON API specific
alternative, `ActionController::API`, which we use instead.

Limits
------

We want all index actions to allow a limit for the number of records to
be returned. Our default for all of the API v1 will be 100. Since this
will be the default for all collection endpoints we use a class
attribute. If a specific resource wants a different limit they can
override it as a configuration option at the start of the class
definition.

Similarly, we want to prevent users from trying to dump the entire
database. That is not performant and on a production system could easy
crash it.

Performance tuning is always an on going task. As such we anticipate
that we may need to low a controller's maximum record limit in the
future. At that point it's possible a client was previously below the
prior limit but is now above the new limit. For such a case, really any
case where the client is above the limit, there are two general
approaches:

1. Error indicating the request in invalid
2. Sliently use the smaller of the max and user provided limits

To be kind to clients we take option two. Last thing we want is an
accidental production fire because a client started receiving unexpected
errors when it was working previously.

Auth
----

I believe all APIs should have some form of authentication; if nothing
else it provides a way to identify and rate limit bad actors.

In this case we are developing message API. The first feature we are
going to work on is direct messages between two users. We want to treat
DMs as private and thus we want to restrict access scopes.

Due to time constraints this does not implement any auth. We could have
added some auth tokens and stored them in plain text, but that's not a
practice I like to perpetuate; even in a demo app. So instead, we going
to use the user ID prepended with "AUTH-PLACEHOLDER-". The intent here
is to be extremely clear that this is a temporary implementation meant
to inform the design of the controllers and nothing else.

We'll make implementing a more secure auth top priority after the POC is
ready.