Use Swagger.io and rswag for API docs
We're picking Swagger.io as it has been around for a while and works
with the openapis.org spec. Swagger.io provides out of the box
formatting, examples, curl commands etc. as part of the documentation.
Additionally, it has an online live editor for free, as well as, an
basic open source editor for the lone developer.

In addition to the API docs, the Swagger.io tool chain takes the OpenAPI
definition and makes it easy to create client libraries and even server
stubs in a variety of languages. And for businesses there's a whole set
of additional tool chains with more advanced features.

We'll be working with Swagger via the `rswag` gem. The gem provides easy
plug-and-play engines to self host Swagger.io docs directly alongside
the app. It also has a test harness so that our request specs generate
the API documentation. It's always good to know that the API
documentation generated matches the code base. It's great for developers
as we just focus on the testing and the docs are easily updated.

For the initial demo POC we aren't going to customize the configuration
at all.