Spike out messages API
The bulk of the controller and views were generated via the Rails
generators.

One notable change is in regards to the `depot` association. The default
generators created `depot_id`. This switches that to be the nested
object `depot` which contains the compound key: `receiver_id` and
`receiver_type`. I'm not happy with this decision, but it was made
quickly due to time constraints.

This primarily tests the happy paths for the `GET` actions. The goal
here was to ensure that some of the documentation was correctly
generated.

There's a lot of work still to do here to flush out the final JSON
payload. Decide how we can filter `messages` by a specific user (or
group in the future) and shore up error responses.